Add helper to infer a CryptoKey acceptable usage based on its type
In order to decide on which usages a generated CryptoKey is valid for,
we needed a way to quickly find out the allowed usages for each key
format. Inside a given algorithm kind (Rsa, ECDSA, ECDH, etc), different
cipher, and whether the key is public or private, allow for different
usages too.

This commit introduces a method holding a table to obtain this
information from an algorithm name, and crypto key type.